Added RTL layout support for Windows.
